move help widget to extensions
