+++ examples/3DTilesDebugWidget/example.html
@@ -25,7 +25,7 @@
         app.addModuleView('about', about);
 
         ////// HELP MODULE
-        const help = new udv.Widgets.HelpWindow(config.helpWindow);
+        const help = new udv.Widgets.Extensions.HelpWindow(config.helpWindow);
         app.addModuleView('help', help);
 
         ////// 3DTILES DEBUG
