+++ examples/CityObjectWidget/example.html
@@ -28,7 +28,7 @@
         app.addModuleView('about', about);
 
         ////// HELP MODULE
-        const help = new udv.Widgets.HelpWindow(config.helpWindow);
+        const help = new udv.Widgets.Extensions.HelpWindow(config.helpWindow);
         app.addModuleView('help', help);
 
         ////// CITY OBJECTS MODULE
