+++ examples/LayerChoiceWidget/example.html
@@ -27,7 +27,7 @@
           app.addModuleView('about', about);
 
           ////// HELP MODULE
-          const help = new udv.Widgets.HelpWindow(config.helpWindow);
+          const help = new udv.Widgets.Extensions.HelpWindow(config.helpWindow);
           app.addModuleView('help', help);
 
           ////// LAYER CHOICE MODULE
