+++ src/Widgets/Extensions/Extensions.js
@@ -15,3 +15,5 @@ export { Debug3DTilesWindow } from './3DTilesDebug/views/3DTilesDebugWindow';
 export { ContributeModule } from './Contribute/ContributeModule';
 
 export { DocumentValidationModule } from './DocumentValidation/DocumentValidationModule';
+
+export { HelpWindow } from './Help/Help';
