+++ src/Widgets/Extensions/Help/Help.js
@@ -1,7 +1,7 @@
 /** @format */
 
 //Components
-import { ModuleView } from '../Components/ModuleView/ModuleView';
+import { ModuleView } from '../../Components/ModuleView/ModuleView';
 
 import jQuery from 'jquery';
